Add 'fairness calculator'
This will be used to keep track of the 'fairness' of allocated
shifts per developer. We know that being a primary on-call person
is more burdensome than, say, being inhours standby secondary,
so we can't go on shift allocation counts alone; we'll want to
allow the concept of some shifts being worth 'more' than others
and thus prevent developers from being over-burdened.