Add integration test
It can be hard to capture all the complexities of a real world
example, and also to have confidence that the script won't crash
on a more varied dataset. Therefore here's a test with real data,
which we should manually sense-check and update whenever we make
a logical change to the rota generator, to ensure that it still
produces balanced output.