Add secondary random sorting in generator
Sort order when fairness is equal is undefined, and appears to behave
differently on OSX and Linux. To stabilise it, I've added a random
factor to the Person model when it's created, based on a singleton
randomiser. We use this as a secondary sort, which avoids the undefined
sort order issue. The randomisation means we can use a fixed seed in
tests but do not have to rely on sorting secondarily by (say) name,
which might result in unfortunate subtle effects on the rota.