Remove 'shadow' role from rota generator - all roles now mandatory
Shadow shifts are no longer coupled to getting Production Admin
access - see https://docs.google.com/document/d/1dv6TKd5UrPGnOFkNRnDZOkZjAn_vr62cb3cVPxl1GqU/edit
and https://github.com/alphagov/govuk-developer-docs/pull/4478

This means that there is no hard requirement to assign engineers
to shadow shifts before joining 2nd line 'for real'. The
prerequisite for being a Secondary on 2nd line is that one has
Production Admin access, and one can gain Production Admin access
outside of 2nd line.

That's not to say that shadow shifts are a thing of the past - it
might be beneficial for folks to take a shadow shift before their
first Secondary shift - but it is no longer mandated and does not
need to be factored in as part of the rota generation. Engineers
can instead arrange with 2nd line leads for a shadow shift.

This change simplifies the rota generator ahead of us making
further improvements (that would otherwise be complicated by the
shadow shift requirement).