Add 'name' method to Person, deriving name from their email
We will likely be using the name in the rota output (for the
2nd line rota spreadsheet) as emails are verbose, and especially
so if we're listing multiple folks for a given week because of
non working days.

Note that I've included an apostrophe in the test email just to
make sure the application doesn't break. I'm not too concerned
about the accuracy of the output - it should probably be "O'Hara"
rather than "O'hara", but then other names don't capitalize after
the apostrophe (e.g. "Lupita Nyong'o"). So rather than bake
complexity into the system, I'm happy with the output for now and
we can revisit it later if needed.
(We don't currently have any apostrophes in any of the engineers
names on the rota).