Change Person class to work with 'forbidden days' rather than weeks
Note that this breaks the RotaGenerator, but that will need such
a rewrite that we'll keep those changes for a commit later in the
PR, to keep things manageable.

Note also that we've had to describe the nature of each 'role'
(in hours vs on call) to avoid having to hardcode unnecessary
business logic in Person.