Change `Person` to work with 'forbidden days' rather than weeks
Note that we've had to describe the nature of each 'role'
(in hours vs on call) to avoid having to hardcode unnecessary
business logic in Person. This helps to keep Person generic,
meaning we can define different role types in future and only have
to do so in the roles config file, without having to edit Person.