Fix rota generator code to work with dates rather than weeks
Note that this removes the "can handle real datasets" test, as
that will require a significant rewrite.
The input data will now be a YAML file (encompassing the arrays of
forbidden in-hours and on-call dates) and we will also have to
generate an array of dates to pass to the rota generator.

The intention is that this test will return in some form, in a
future commit in this same PR / body of work.