Add DataProcessor class
The intention is for this to be the single place that maps
responses/people data to the format required by the rota
generator. We'll remove the resulting redundant code in the next
couple of commits.

Note that the Google Form template lists working days under their
shortname (e.g. "Fri"), for brevity, so we convert these to their
full name to remove ambiguity from future uses in this repo.