Add `to_h` method on Person
This implementation is based on
https://dev.to/ayushn21/how-to-generate-yaml-from-ruby-objects-without-type-annotations-4fli

We will soon need to output an array of Person to a YAML file, so
that we can decouple the "data fetching" and "rota generating"
steps of the rota generator. We will use the new `Person.to_h`
method to describe the Person's YAML representation.