Assume an engineer is available if they haven't provided response
This might seem a harsh stance, but an engineer shouldn't be opted
out of 2nd line just because they didn't go to the effort of
providing a form response.

That said, the reason for this commit is more to do with us
wanting to make a note of all engineers (whether or not they've
filled in the availability survey) in the easiest way possible.