Sort assigned shifts by date
This will be needed in the next commit.

Previously, we'd made an assumption that all 'shifts' are assigned
in date order, but that will not necessarily always be the case.
Sometimes we need to 'find cover' for folks who can't manage an
entire shift, and this will likely be implemented in a 'cleanup'
phase, causing the shifts to be assigned out of date order.