Encapsulate 'roles' in config file
We'll soon add additional metadata to each role (denoting whether
or not the role is in-hours or out-of-hours, in a more explicit
way than trying to derive it from the name). When we have that
extra metadata, it will be extremely handy to be able to read from
a config file rather than try to change the shape of the data
that's being passed into Person.rb, for example.