Add 'DataProcessor' class
The intention is for the DataProcessor class to be the only place
we need to update whenever there is a change in the structure of
the upstream data (Technical Support sheet and Availability /
Responses sheet). Over the next few commits, we'll delete a lot
of code from elsewhere that is made redundant as a result.

Note that the `combine_raw` method is not intended to be called
from upstream: it is only exposed for testing. In the next commit
we'll write the `combine` method that takes CSV file inputs and
outputs a YML file, and uses `combine_raw` under the hood.