Add 'DataProcessor' class & 'create_people_from_csv_data' method
The intention is for the DataProcessor class to be the only place
we need to update whenever there is a change in the structure of
the upstream data (Technical Support sheet and Availability /
Responses sheet). Over the next few commits, we'll delete a lot
of code from elsewhere that is made redundant as a result.