Remove Rota Generator v1
The underlying data model is changing from 'weekly' (with a lot of
assumptions about the state of the world) to 'daily', so that we
can account for non-working days and the like.

Early attempts at changing this behaviour whilst trying to retain
a lot of the original code and tests were more complex than they
were worth. Better to delete what we will no longer need, and
implement a new rota generator from scratch.