Add `to_h` method on Person
This implementation is based on
https://dev.to/ayushn21/how-to-generate-yaml-from-ruby-objects-without-type-annotations-4fli

We will soon need to output an array of Person to a YAML file, so
that we can decouple the "data fetching" and "rota generating"
steps of the rota generator. We will also populate instances of
Person from that YAML file (in the rota generation step).

We will use the new `Person.to_h` method to describe the Person's
YAML representation. We've also had to make 'assigned_shifts' a
keyword parameter, as there may well be the presence of an
'assigned_shifts' property in the YAML file. It therefore needs to
be a keyword argument of the class, to not blow up.