Write RotaPresenter for transforming YML output to CSV
Keeping this separate from the RotaGenerator allows us to more
easily test it (we can change the underlying algorithm that
generates the rota, and the tests for RotaPresenter will be
unaffected, as they _only_ look at the generated_rota.yml, which
will remain constant in the fixtures).

It also provides room for a "output this as a 'summarised' weekly
CSV" - watch this space!