Write `RotaPresenter` class
The underlying data now uses individual dates rather than weeks,
but we want to output a 'summarised' CSV that consolidates the
data into weeks at a time. This is implemented in the
`to_csv_weekly` method. The `to_csv_daily` method is provided
for completion.

`to_yaml` stores the underlying data in a lossless way (retaining
availability, team name, and so on), so that it can be fed back
into a rota algorithm to generate a different rota later if needed.

The `initialize` method can either initialise `RotaPresenter`
from said YAML file, or directly from the data that populates
the YAML file. The latter is useful for testing and the former
is likely more useful for the rota generation script itself.