Write Weekly algorithm
In earlier iterations of this PR, we had a `RotaGenerator` class
that took an arbitrary 'algorithm'. For example, it could take an
algorithm that outputs a rota that has a different engineer each
day rather than each week. I therefore opted to create an
`algorithms` directory and start putting the algorithms in there.

The `RotaGenerator` class ended up just being a wrapper for the
algorithm, passing parameters down a level and offering no value
in itself, so was later rebased out of the PR. The
`Algorithm::Weekly` that remains is useable in isolation and will
be called from the `bin/generate_rota.rb` script later.

Note that this algorithm fills roles first, rather than weeks -
the theory being that the roles themselves will be more evenly
distributed as a result. We also process the roles in order of
on call followed by in hours, leading to a more even distribution
because otherwise the pool of available on-call devs is reduced
by the time we come to fill the on-call slots.