Remove Randomiser from Person class
This was only used for an end-to-end generator test that has now
been removed.

The intention is that the new set of tests for 'v2' of the
generator will be better implemented such that we won't need an
end-to-end like the one that has just been removed, and can
reliably run the tests on CI without the randomiser.