Update the roles weighting
With a bit of trial and error, this tends to lead to fairer
outcomes. It seems odd to equate a standby in-hours shift as
almost the same value as a primary on-call shift, but that
_tends_ to mean folks are assigned to multiple in-hours
shifts, and this is unfair in a different way (burden on the
engineer's team). Whereas if, by on large, everyone is assigned
to no more than 2 shifts in a 13 week period, it's generally
fairer.