Add 'pagerduty_user_id' getter/setter to Person
We'll need this for the Pagerduty synchronisaiton script.

We're adding this as a getter/setter rather than expecting it as
a parameter to `initialize`, as we'll need to cross-reference the
person's `name` against the names in PagerDuty, so we won't
necessarily know the Pagerduty user ID at the time of initialising
the `Person` (and we need to initialise `Person` before we can
call `name` on it).