Fix grouping of shifts when someone has a non working Monday
The `name_of_person_with_most_shifts` logic was broken - it was
simply returning the first person chronologically who is assigned
a shift for a given role. What we actually want it to do is
return the name of the person with the _most_ shifts for a given
role, so that we can show their name by default and show fewer
overrides.

Example before:

```
inhours_standby
David Mays (Minno Dang on 02/07/2024, Minno Dang on 03/07/2024, Minno Dang on 04/07/2024, Minno Dang on 05/07/2024)
```

Example after:

```
inhours_standby
Minno Dang (David Mays on 01/07/2024)
```