Assign the on-call devs to in-hours shifts on bank holidays
How we assign bank holiday weekends is up for debate - but what
is obvious is that we shouldn't expect the assigned in-hours folks
to log on for a 9:30am-5:30pm shift on the bank holiday. It is far
less surprising for the assigned on-call person to be the one
assigned the daytime shift in PagerDuty.