Apply overrides until 9:30am the day after the final day of rota
The Pagerduty sync script was only applying overrides until 00:00
on the last day of the rota. But the rota the next quarter starts
at 9:30am, meaning we had an unaccounted 00:00-09:30 period where
PagerDuty assigned someone at random.

The reason the override finished at 0:00 was because the PagerDuty
method `pd.schedule` returns all of the shifts for the given start
and end dates, and we then blindly reuse the 'start' and 'end'
times when overriding via `pd.create_override`. Because we'd only
given PagerDuty a range of, say, 1st July to 29 September, for the
last day it would return an on-call shift of `17:30-00:00`, cutting
off the `00:00-9:30` portion, since that would spill into 30
September.

We now explicitly request up to 09:30 the day after the last day
of the rota, meaning the last shift now returns `17:30-9:30`, and
meaning we override the entire shift.

This only affects the on-call shifts - in-hours shifts continue to
end at 17:30 (automatically accounted for by the `start` and `end`
times we pass to PagerDuty).

Here are the before/after screenshots of the updated script fixing
up the midnight-9:30 slots.

| Before | After |
|-------|-------|
| <img alt="primary before" src="https://github.com/alphagov/govuk-rota-generator/assets/5111927/83043ece-09cd-448d-8c19-bf423cc8c651" width="400px"> | <img alt="primary after" src="https://github.com/alphagov/govuk-rota-generator/assets/5111927/39671c80-83ec-4b47-b711-a9d70e457069" width="400px"> |
| Primary | |
| <img alt="secondary before" src="https://github.com/alphagov/govuk-rota-generator/assets/5111927/efef4e11-be15-4b93-9899-136036970441" width="400px"> | <img alt="secondary after" src="https://github.com/alphagov/govuk-rota-generator/assets/5111927/af60099b-bd59-4722-9c0c-cb68a7f99a5f" width="400px"> |
| Secondary | |

Trello: https://trello.com/c/6osPAiWi/221-ensure-generated-rotas-finish-at-930am-rather-than-midnight