Document fixing of 'failed to assign' bug
We used to have to manually apply overrides in the UI whenever a
shift spanned multiple slots (e.g. 17:30 Friday until 09:30 Monday
as opposed to something simple like 17:30 Thursday until 09:30
Friday).

Since the inclusion of the 'end' time in evaluating which shifts
to include in `pagerduty_shifts_to_override`, we now have greater
granularity and control over shift overrides, and we were seeing
cases where, for example:

- Someone should be scheduled 17:30 Friday - 09:30 Monday
- They're already scheduled for 09:30 Saturday - 09:30 Monday and
  just need the 17:30 Friday - 09:30 Saturday slot assigning.
- The script detects the 17:30 Friday - 09:30 Saturday slot needs
  assigning and offers to fix it up...
- ...but then also unnecessary prompts to fix up the 09:30
  Saturday - 09:30 Sunday, and 09:30 Sunday - 09:30 Monday slots,
  even though they're already correctly assigned.

So now we check if the person being overridden is the same person,
and if so, skip over it.