Move 'shifts assigned to wrong person' logic into PagerdutyClient
Again, this logic is important and should be tested. Whilst at
this point we're not interfacing with PagerDuty API, we are
interfacing with the _response_ from an earlier PagerDuty API
call, so it feels a reasonable place to bundle the logic into.

Defines new `shifts_assigned_to_wrong_person` method, which is
a much clearer naming and helps to differentiate between this
logic and the `pagerduty_shifts_to_override` logic later in the
bin script.