Move 'get shifts within start/end times' logic into PagerdutyClient
This moves the last of the start/end datetime logic into somewhere
where it can be unit tested. New method is named
`shifts_within_timespan`, removing the last of the 'pagerduty
shifts to override' style naming which was always a bit murky.
This method clearly just returns the subset of PagerDuty shifts
that fall between the given start/end timestamps.