Document the cronjob and set it up to run hourly during the day
We want to accommodate any changes that may occur during the day,
e.g. a shift swap on the Google Sheet. This should automatically
be picked up on the hour, when the cron job runs. No changes
should happen outside of working hours, so 8am-6pm seems a
reasonable time frame.