Avoid scheduling back-to-back shifts
The rota generator, whilst doing a pretty good job at evenly
distributing shifts, was allocating multiple back-to-back shifts.
E.g. Jane Doe down for out-of-hours on-call two weeks in a row.

This was especially true for on-call secondary one week followed
by on-call primary the next week, which stems from how the rota
is allocated:

1. Allocate all on-call primary shifts first, then
2. Allocate all on-call secondary shifts,
3. (then the in-hours shifts)

One can imagine how Jane Doe may be allocated an on-call primary
shift in Week 2, and then allocated an on-call secondary shift
in Week 1, where existing checks _may_ have caught a "has this
person been allocated a shift in the previous week?" kind of
situation, but won't have caught a "is this person scheduled for
a shift _next week_?" situation.

This does unfortunately make the code that bit more horrible to
maintain. The generator would be ripe for a bit of a refactor
when we're not so close to the wire in terms of getting a rota
out.

Trello: https://trello.com/c/y2u2lFfq/300-avoid-scheduling-back-to-back-shifts