Fix PagerDuty synchronisation logic
The synchronisation script was consistently failing for bank holiday
adjacent shifts:

```
Overriding 6 individual shifts in PagerDuty...
Warning: failed to assign Bruce Bolt to the oncall_secondary role from 2025-05-05T17:30:00+01:00 to 2025-05-06T09:30:00+01:00. You'll need to apply this manually in the PagerDuty UI.
Warning: failed to assign Ken Tsang to the oncall_secondary role from 2025-04-16T17:30:00+01:00 to 2025-04-17T09:30:00+01:00. You'll need to apply this manually in the PagerDuty UI.
Warning: failed to assign Ken Tsang to the oncall_secondary role from 2025-04-17T17:30:00+01:00 to 2025-04-18T09:30:00+01:00. You'll need to apply this manually in the PagerDuty UI.
Warning: failed to assign Ken Tsang to the oncall_secondary role from 2025-04-18T17:30:00+01:00 to 2025-04-21T09:30:00+01:00. You'll need to apply this manually in the PagerDuty UI.
Warning: failed to assign Kevin Dew to the oncall_secondary role from 2025-04-21T17:30:00+01:00 to 2025-04-22T09:30:00+01:00. You'll need to apply this manually in the PagerDuty UI.
Warning: failed to assign Keith Lawrence to the oncall_secondary role from 2025-05-26T17:30:00+01:00 to 2025-05-27T09:30:00+01:00. You'll need to apply this manually in the PagerDuty UI.
Finished overriding oncall_secondary shifts.
```

This was because the script was expecting each shift to neatly
line up in 9:30-17:30 and 17:30-9:30 batches. When the out-of-hours
on-call person now also has a 9:30-17:30 shift (representing
'in-hours' on, say, a bank holiday Monday) then the timestamps
didn't match exactly and we had this warning being output to the
console, even though, in reality, the shifts had been successfully
applied.

We also had a rather nasty loop-inside-a-loop, where we first of
all try to establish which shifts need overriding and then we also
try to establish who the current 'holders' of the shift are. This
led to a weird state of affairs where we'd have, say, 3 PagerDuty
shifts that need overriding, but on closer inspection maybe two of
them are already assigned to the right person, so the 'actual'
number of shifts that need overriding in that case is 1.

This commit does away with that needless complication, and just
throws the shift start/end dates at PagerDuty. We do fetch the
names of the corresponding 'holders' of the shift(s) for debug
purposes, but it's not longer central to the logic of how it all
works.