Fix the "Overriding n shifts in PagerDuty..." message
This would sometimes display the wrong number, if a shift was
detected that has the wrong person assigned to it, but that shift
is also in the past. We'd display "Overriding 1 shifts in PagerDuty"
and then immediately call `next` on the shift that is in the past,
thereby not performing any override at all.